% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{CommandBar.shinyInput}
\alias{CommandBar.shinyInput}
\title{CommandBar.shinyInput}
\usage{
CommandBar.shinyInput(inputId, ..., itemValueGetter = function(el) el$key)
}
\arguments{
\item{inputId}{Input name}

\item{...}{Arguments passed to CommandBar.}

\item{itemValueGetter}{A function that takes a CommandBarItem and returns a value to be sent to Shiny.
By default it returns \code{key} of the item.}
}
\value{
A CommandBar component.
}
\description{
CommandBar extension that sends values of clicked CommandBarItems
}
\seealso{
CommandBarItem
}
