% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{CommandBarItem}
\alias{CommandBarItem}
\title{Command bar item}
\usage{
CommandBarItem(
  key,
  text,
  onClick = setInputValue(inputId = key, value = 0, event = TRUE),
  ...
)
}
\arguments{
\item{key}{Key of the item.}

\item{text}{Text to be displayed on the menu.}

\item{onClick}{A JS function that runs on item click. By default it sends input value to \code{input[[key]]}.
If used within \code{CommandBar.shinyInput}, it will send the value to the input ID specified
in \code{inputId} argument of \code{CommandBar.shinyInput}.}

\item{...}{Additional props to pass to CommandBarItem.}
}
\value{
Item suitable for use in the \code{CommandBar} and \code{CommandBar.shinyInput}.
}
\description{
Helper function for constructing items for \code{CommandBar} and \code{CommandBar.shinyInput}.
}
\seealso{
CommandBar
}
