% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{doc}
\alias{doc}
\title{Ellipsis documentation template}
\arguments{
\item{...}{Props to pass to \code{shiny.react::reactElement}. See allowed props listed below}
}
\description{
Ellipsis documentation template
}
