% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{ExtendedPeoplePicker}
\alias{ExtendedPeoplePicker}
\title{ExtendedPeoplePicker}
\arguments{
\item{...}{Props to pass to \code{shiny.react::reactElement}. See allowed props listed below in \bold{Details} section}
}
\description{
For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/ExtendedPeoplePicker}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
