% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluent_people}
\alias{fluent_people}
\title{A dataset of sample people based on Fluent UI examples}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 11 columns.
}
\source{
\url{https://developer.microsoft.com/en-us/fluentui#/controls/web/peoplepicker}
}
\usage{
fluent_people
}
\description{
A dataset of sample people based on Fluent UI examples
}
\keyword{datasets}
