% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluent_sales_deals}
\alias{fluent_sales_deals}
\title{A randomly generated dataset of imaginary sales deals}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92 rows and 30 columns.
}
\usage{
fluent_sales_deals
}
\description{
Sales deals to Top 10 companies from the Fortune 500 dataset (located at
\url{https://hifld-geoplatform.opendata.arcgis.com/datasets/fortune-500-corporate-headquarters})
are randomly generated for each person in \code{fluent_people}.
}
\keyword{datasets}
