% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{inputDoc}
\alias{inputDoc}
\title{Input docs template}
\arguments{
\item{inputId}{Id of the component}

\item{value}{Value}

\item{session}{Session}

\item{...}{Props to be passed}
}
\description{
Input docs template
}
