% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{MarqueeSelection}
\alias{MarqueeSelection}
\title{MarqueeSelection}
\usage{
MarqueeSelection(...)
}
\arguments{
\item{className}{\code{string} \cr Additional CSS class(es) to apply to the MarqueeSelection.}

\item{componentRef}{\verb{IRefObject<IMarqueeSelection>} \cr Optional callback to access the IMarqueeSelection interface. Use this instead of ref for accessing the public methods and properties of the component.}

\item{isDraggingConstrainedToRoot}{\code{boolean} \cr Optional flag to restrict the drag rect to the root element, instead of allowing the drag rect to start outside of the root element boundaries.}

\item{isEnabled}{\code{boolean} \cr Optional flag to control the enabled state of marquee selection. This allows you to render it and have events all ready to go, but conditionally disable it. That way transitioning between enabled/disabled generate no difference in the DOM.}

\item{onShouldStartSelection}{\verb{(ev: MouseEvent) => boolean} \cr Optional callback that is called, when the mouse down event occurs, in order to determine if we should start a marquee selection. If true is returned, we will cancel the mousedown event to prevent upstream mousedown handlers from executing.}

\item{rootProps}{\verb{React.HTMLAttributes<HTMLDivElement>} \cr Optional props to mix into the root DIV element.}

\item{selection}{\code{ISelection} \cr The selection object to interact with when updating selection changes.}

\item{styles}{\verb{IStyleFunction<IMarqueeSelectionStyleProps, IMarqueeSelectionStyles>} \cr Call to provide customized styling that will layer on top of the variant rules.}

\item{theme}{\code{ITheme} \cr Theme (provided through customization.)}
}
\description{
The MarqueeSelection component provides a service which allows the user to drag a rectangle to be drawn around
items to select them. This works in conjunction with a selection object, which can be used to generically store selection state, separate from a component that consumes the state.

MarqueeSelection also works in conjunction with the AutoScroll utility to automatically scroll the container when we drag a rectangle within the vicinity of the edges.

When a selection rectangle is dragged, we look for elements with the \strong{data-selection-index} attribute populated. We get these elements' boundingClientRects and compare them with the root's rect to determine selection state. We update the selection state appropriately.

In virtualization cases where items that were once selected are dematerialized, we will keep the item in its
previous state until we know definitively if it's on/off. (In other words, this works with List.)

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/MarqueeSelection}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
