% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{parseTheme}
\alias{parseTheme}
\title{parseTheme}
\usage{
parseTheme(json)
}
\arguments{
\item{json}{Either a JSON string or a path to JSON file containing the theme created in Theme Designer}
}
\description{
Reads a theme JSON generated by Theme Designer: \url{https://fabricweb.z5.web.core.windows.net/pr-deploy-site/refs/heads/master/theming-designer/}
and parses it to an object digestable by \code{ThemeProvider}
}
\seealso{
\code{\link[=ThemeProvider]{ThemeProvider()}} for usage of this function
}
