% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{pickers}
\alias{pickers}
\title{pickers}
\description{
Pickers are used to select one or more items, such as tags or files, from a large list.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/pickers}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\section{Best practices}{
\subsection{Layout}{
\itemize{
\item Use a picker to quickly search for a few tags or files.
\item Use a picker to manage a group of tags or files.
}
}
}

