% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{Popup}
\alias{Popup}
\title{Popup}
\arguments{
\item{ariaDescribedBy}{\code{string} \cr Defines the element id referencing the element containing the description for the popup.}

\item{ariaLabel}{\code{string} \cr Accessible label text for the popup.}

\item{ariaLabelledBy}{\code{string} \cr Defines the element id referencing the element containing label text for popup.}

\item{className}{\code{string} \cr Optional class name for the root popup div.}

\item{onDismiss}{\verb{(ev?: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>) => any} \cr A callback function for when the popup is dismissed from the close button or light dismiss. If provided, will handle escape keypresses and call this. The event will be stopped/canceled.}

\item{onRestoreFocus}{\verb{(options: \{ originalElement?: HTMLElement | Window; containsFocus: boolean; \}) => void} \cr Called when the component is unmounting, and focus needs to be restored. Argument passed down contains two variables, the element that the underlying popup believes focus should go to * and whether or not the popup currently contains focus. If this is provided, focus will not be restored automatically, you'll need to call originalElement.focus()}

\item{role}{\code{string} \cr Aria role for popup}

\item{shouldRestoreFocus}{\code{boolean} \cr If true, when this component is unmounted, focus will be restored to the element that had focus when the component first mounted.}
}
\description{
For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/Popup}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
