% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run example shiny.fluent apps.}
\usage{
run_example(
  example = NA,
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = c("auto", "normal", "showcase")
)
}
\arguments{
\item{example}{Example to run. \code{NA} to list the examples.}

\item{launch.browser}{Whether to open the app in a browser}

\item{host}{The IPv4 address to listen on.}

\item{display.mode}{Display mode for the app.}
}
\description{
Based on shiny::runExample, and takes the same arguments.
}
