% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{runExample}
\alias{runExample}
\title{Run example shiny.fluent apps.}
\usage{
runExample(
  example = NA,
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = c("auto", "normal", "showcase")
)
}
\arguments{
\item{example}{Example to run. \code{NA} to list the examples.}

\item{port}{The TCP port that the application should listen on}

\item{launch.browser}{Whether to open the app in a browser}

\item{host}{The IPv4 address to listen on.}

\item{display.mode}{Display mode for the app.}
}
\value{
This function normally does not return;
interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\description{
Based on shiny::runExample, and takes the same arguments.
}
