% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{SelectedPeopleList}
\alias{SelectedPeopleList}
\title{SelectedPeopleList}
\arguments{
\item{...}{Props to pass to \code{shiny.react::reactElement}. See allowed props listed below in \bold{Details} section}
}
\description{
For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/SelectedPeopleList}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\details{
\itemize{
\item {\bold{ blockRecipientRemoval }} {\code{boolean} \cr }
\item {\bold{ canExpand }} {\code{boolean} \cr }
\item {\bold{ isEditing }} {\code{boolean} \cr }
\item {\bold{ isValid }} {\code{boolean} \cr }
\item {\bold{ key }} {\code{React.Key} \cr }
\item {\bold{ shouldBlockSelection }} {\code{boolean} \cr }
\item {\bold{ onExpandItem }} {\verb{() => void} \cr }
\item {\bold{ renderPersonaCoin }} {\verb{IRenderFunction<IPersonaProps>} \cr }
\item {\bold{ renderPrimaryText }} {\verb{IRenderFunction<IPersonaProps>} \cr }
\item {\bold{ copyMenuItemText }} {\code{string} \cr }
\item {\bold{ editMenuItemText }} {\code{string} \cr }
\item {\bold{ floatingPickerProps }} {\verb{IBaseFloatingPickerProps<IPersonaProps>} \cr }
\item {\bold{ getEditingItemText }} {\verb{(item: IExtendedPersonaProps) => string} \cr }
\item {\bold{ onExpandGroup }} {\verb{(item: IExtendedPersonaProps) => void} \cr }
\item {\bold{ onRenderFloatingPicker }} {\verb{React.ComponentType<IBaseFloatingPickerProps<IPersonaProps>>} \cr }
\item {\bold{ removeMenuItemText }} {\code{string} \cr }
}
}
