% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Separator}
\alias{Separator}
\title{Separator}
\usage{
Separator(...)
}
\arguments{
\item{alignContent}{\code{'start' | 'center' | 'end'} \cr Where the content should be aligned in the separator.}

\item{styles}{\verb{IStyleFunctionOrObject<ISeparatorStyleProps, ISeparatorStyles>} \cr Call to provide customized styling that will layer on top of the variant rules.}

\item{theme}{\code{ITheme} \cr Theme (provided through customization.)}

\item{vertical}{\code{boolean} \cr Whether the element is a vertical separator.}
}
\description{
A separator visually separates content into groups.

You can render content in the separator by specifying the component's children.
The component's children can be plain text or a component like Icon.
The content is center-aligned by default.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/Separator}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\examples{
library(shiny.fluent)

if (interactive()) {
  shinyApp(
    ui = withReact(
      Separator("Text")
    ),
    server = function(input, output) {
    }
  )
}
}
