% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{shinyInput}
\alias{shinyInput}
\title{Documentation template for components with \code{.shinyInput} wrappers}
\arguments{
\item{inputId}{ID of the component.}

\item{value}{Starting value.}

\item{session}{Object passed as the \code{session} argument to Shiny server.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Documentation template for components with \code{.shinyInput} wrappers
}
\keyword{internal}
