% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{Themes}
\alias{Themes}
\title{Fluent UI themes}
\description{
Themes is not a component, but a way to override the default theme in the app.
See https://developer.microsoft.com/en-us/fluentui#/controls/web/themes for details.
}
\examples{
library(shiny.fluent)

script <- tags$script(HTML("
Fluent = window['shiny.fluent']
Fluent.loadTheme({
  defaultFontStyle: { fontFamily: 'Times New Roman', fontWeight: 'regular' }
})
"))

if (interactive()) {
  shinyApp(
    ui = tagList(
      script,
      withReact(
        Label('The default theme font has been changed to Times New Roman in this example')
      )
    ),
    server = function(input, output) {
    })
}
}
