% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Breadcrumb}
\alias{Breadcrumb}
\title{Breadcrumb}
\usage{
Breadcrumb(...)
}
\arguments{
\item{ariaLabel}{\code{string} \cr Aria label for the root element of the breadcrumb (which is a navigation landmark).}

\item{className}{\code{string} \cr Optional class for the root breadcrumb element.}

\item{componentRef}{\verb{IRefObject<IBreadcrumb>} \cr Optional callback to access the IBreadcrumb interface. Use this instead of ref for accessing the public methods and properties of the component.}

\item{dividerAs}{\verb{IComponentAs<IDividerAsProps>} \cr Render a custom divider in place of the default chevron \code{>}}

\item{focusZoneProps}{\code{IFocusZoneProps} \cr Extra props for the root FocusZone.}

\item{items}{\code{IBreadcrumbItem[]} \cr Collection of breadcrumbs to render}

\item{maxDisplayedItems}{\code{number} \cr The maximum number of breadcrumbs to display before coalescing. If not specified, all breadcrumbs will be rendered.}

\item{onGrowData}{\verb{(data: IBreadcrumbData) => IBreadcrumbData | undefined} \cr Method that determines how to group the length of the breadcrumb. Return undefined to never increase breadcrumb length.}

\item{onReduceData}{\verb{(data: IBreadcrumbData) => IBreadcrumbData | undefined} \cr Method that determines how to reduce the length of the breadcrumb. Return undefined to never reduce breadcrumb length.}

\item{onRenderItem}{\verb{IRenderFunction<IBreadcrumbItem>} \cr Custom render function for each breadcrumb item.}

\item{onRenderOverflowIcon}{\verb{IRenderFunction<IButtonProps>} \cr Render a custom overflow icon in place of the default icon \code{...}}

\item{overflowAriaLabel}{\code{string} \cr Aria label for the overflow button.}

\item{overflowIndex}{\code{number} \cr Optional index where overflow items will be collapsed. Defaults to 0.}

\item{styles}{\verb{IStyleFunctionOrObject<IBreadcrumbStyleProps, IBreadcrumbStyles>} \cr}

\item{theme}{\code{ITheme} \cr}

\item{tooltipHostProps}{\code{ITooltipHostProps} \cr Extra props for the TooltipHost which wraps each breadcrumb item.}

\item{item}{\code{IBreadcrumbItem} \cr Breadcrumb item to left of the divider to be passed for custom rendering. For overflowed items, it will be last item in the list.}
}
\description{
Breadcrumbs should be used as a navigational aid in your app or site. They indicate the current page’s location within a hierarchy and help the user understand where they are in relation to the rest of that hierarchy. They also afford one-click access to higher levels of that hierarchy.

Breadcrumbs are typically placed, in horizontal form, under the masthead or navigation of an experience, above the primary content area.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/Breadcrumb}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\examples{
library(shiny.fluent)

if (interactive()) {
  items <- list(
    list(text = "Files", key = "Files", href = "#/page"),
    list(text = "Folder 1", key = "f1", href = "#/page"),
    list(text = "Folder 2", key = "f2", href = "#/page"),
    list(text = "Folder 3", key = "f3", href = "#/page"),
    list(text = "Folder 4 (non-clickable)", key = "f4"),
    list(text = "Folder 5", key = "f5", href = "#/page", isCurrentItem = TRUE)
  )

  shinyApp(
    ui = withReact(
      Breadcrumb(
        items = items,
        maxDisplayedItems = 3,
        ariaLabel = "Breadcrumb with items rendered as links",
        overflowAriaLabel = "More links"
      )
    ),
    server = function(input, output) {
    }
  )
}
}
