% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R,
%   R/inputs.R
\name{Calendar}
\alias{Calendar}
\alias{Calendar.shinyInput}
\alias{updateCalendar.shinyInput}
\title{Calendar}
\usage{
Calendar(...)

Calendar.shinyInput(inputId, ..., value = defaultValue)

updateCalendar.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{value}{\code{Date} \cr Default value of the Calendar, if any}

\item{allFocusable}{\code{boolean} \cr Allows all dates and buttons to be focused, including disabled ones}

\item{autoNavigateOnSelection}{\code{boolean} \cr Whether the month view should automatically navigate to the next or previous date range depending on the selected date. If this property is set to true and the currently displayed month is March 2017, if the user clicks on a day outside the month, i.e., April 1st, the picker will automatically navigate to the month of April.}

\item{className}{\code{string} \cr Optional class name to add to the root element.}

\item{componentRef}{\verb{IRefObject<ICalendar>} \cr Optional callback to access the ICalendar interface. Use this instead of ref for accessing the public methods and properties of the component.}

\item{dateRangeType}{\code{DateRangeType} \cr The date range type indicating how many days should be selected as the user selects days}

\item{dateTimeFormatter}{\code{ICalendarFormatDateCallbacks} \cr Apply additional formating to dates, for example localized date formatting.}

\item{firstDayOfWeek}{\code{DayOfWeek} \cr The first day of the week for your locale.}

\item{firstWeekOfYear}{\code{FirstWeekOfYear} \cr Defines when the first week of the year should start, FirstWeekOfYear.FirstDay, FirstWeekOfYear.FirstFullWeek or FirstWeekOfYear.FirstFourDayWeek are the possible values}

\item{highlightCurrentMonth}{\code{boolean} \cr Whether the month picker should highlight the current month}

\item{highlightSelectedMonth}{\code{boolean} \cr Whether the month picker should highlight the selected month}

\item{isDayPickerVisible}{\code{boolean} \cr Whether the day picker is shown beside the month picker or hidden.}

\item{isMonthPickerVisible}{\code{boolean} \cr Whether the month picker is shown beside the day picker or hidden.}

\item{maxDate}{\code{Date} \cr If set the Calendar will not allow navigation to or selection of a date later than this value.}

\item{minDate}{\code{Date} \cr If set the Calendar will not allow navigation to or selection of a date earlier than this value.}

\item{navigationIcons}{\code{ICalendarIconStrings} \cr Customize navigation icons using ICalendarIconStrings}

\item{onDismiss}{\verb{() => void} \cr Callback issued when calendar is closed}

\item{onSelectDate}{\verb{(date: Date, selectedDateRangeArray?: Date[]) => void} \cr Callback issued when a date is selected}

\item{restrictedDates}{\code{Date[]} \cr If set the Calendar will not allow selection of dates in this array.}

\item{selectDateOnClick}{\code{boolean} \cr When clicking on "Today", select the date and close the calendar.}

\item{shouldFocusOnMount}{\code{boolean} \cr This property has been removed at 0.80.0 in place of the \code{focus} method, to be removed \@ 1.0.0.}

\item{showCloseButton}{\code{boolean} \cr Whether the close button should be shown or not}

\item{showGoToToday}{\code{boolean} \cr Whether the "Go to today" link should be shown or not}

\item{showMonthPickerAsOverlay}{\code{boolean} \cr Show month picker on top of date picker when visible.}

\item{showSixWeeksByDefault}{\code{boolean} \cr Whether the calendar should show 6 weeks by default.}

\item{showWeekNumbers}{\code{boolean} \cr Whether the calendar should show the week number (weeks 1 to 53) before each week row}

\item{strings}{\code{ICalendarStrings | null} \cr Localized strings to use in the Calendar}

\item{today}{\code{Date} \cr Value of today. If null, current time in client machine will be used.}

\item{workWeekDays}{\code{DayOfWeek[]} \cr The days that are selectable when dateRangeType is WorkWeek. If dateRangeType is not WorkWeek this property does nothing.}

\item{yearPickerHidden}{\code{boolean} \cr Whether the year picker is enabled}
}
\description{
The calendar control lets people select and view a single date or a range of dates in their calendar. It’s made up of 3 separate views: the month view, year view, and decade view.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/Calendar}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\section{Best practices}{
\subsection{Layout}{
\itemize{
\item Don’t break the control apart.
\item Include an up and down arrow for navigating between time ranges and a chevron to make the calendar collapsible.
}
}

\subsection{Content}{
\itemize{
\item Use the following format for dates: month, day, year, as in July 31, 2016. When space is limited, use numbers and slashes for dates if the code supports that format and automatically displays the appropriate date format for different locales. For example, 2/16/19.
\item Don't use ordinal numbers (such as 1st, 12th, or 23rd) to indicate a date.
}
}
}

\examples{
library(shiny.fluent)

if (interactive()) {
  shinyApp(
    ui = div(
      Calendar.shinyInput("calendar", "2020-06-25T22:00:00.000Z"),
      textOutput("calendarValue")
    ),
    server = function(input, output) {
      output$calendarValue <- renderText({
        sprintf("Value: \%s", input$calendar)
      })
    }
  )
}
}
