% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{CommandBarItem}
\alias{CommandBarItem}
\title{Command bar item}
\usage{
CommandBarItem(text, icon = NULL, subitems = NULL, ...)
}
\arguments{
\item{text}{Text to be displayed on the menu.}

\item{icon}{Optional name of an icon.}

\item{subitems}{Optional list of CommandBar items.}

\item{...}{Additional props to pass to CommandBarItem.}
}
\value{
Item suitable for use in the CommandBar.
}
\description{
Helper function for constructing items for \code{CommandBar}.
}
\details{
CommandBar expects items definition as a nested structure, which gets lengthy and verbose.
This function helps makes this definition shorter. Returns a list with all arguments passed through,
except for \code{text}, \code{icon} (which will inserted as proper \code{iconProps}) and \code{subitems} (which will be inserted as
proper \code{subMenuProps}).
}
\seealso{
CommandBar
}
