% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{CommandBarItem}
\alias{CommandBarItem}
\title{Helper function for constructing items for \code{CommandBar}}
\usage{
CommandBarItem(text, icon = NULL, subitems = NULL, ...)
}
\arguments{
\item{text}{text to be displayed on the menu}

\item{icon}{optional name of an icon}

\item{subitems}{optional list of CommandBar items}
}
\description{
CommandBar expects items definition as a nested structure, which gets lengthy and verbose.
This function helps makes this definition shorter. Returns a list with all arguments passed through,
except for \code{text}, \code{icon} (which will inserted as proper \code{iconProps}) and \code{subitems} (which will be inserted as
proper \code{subMenuProps})
}
\seealso{
CommandBar
}
