% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{DetailsList}
\alias{DetailsList}
\title{DetailsList}
\usage{
DetailsList(...)
}
\arguments{
\item{onColumnDragEnd}{\verb{(props: \{ dropLocation?: ColumnDragEndLocation; \}, event: MouseEvent) => void} \cr Callback to notify the column dragEnd event to List Need this to check whether the dragEnd has happened on corresponding list or outside of the list}

\item{column}{\code{IColumn} \cr The column definition for the component instance.}

\item{columnIndex}{\code{number} \cr The column index for the component instance.}

\item{isDraggable}{\code{boolean} \cr Whether or not the column can be re-ordered via drag and drop.}

\item{isDropped}{\code{boolean} \cr Whether or not the column has been dropped via drag and drop.}

\item{parentId}{\code{string} \cr Parent ID used for accessibility label(s).}

\item{setDraggedItemIndex}{\verb{(itemIndex: number) => void} \cr}

\item{updateDragInfo}{\verb{(props: \{ itemIndex: number; \}, event?: MouseEvent) => void} \cr Callback on drag and drop event.}

\item{onRenderFooter}{\verb{IRenderFunction<IDetailsGroupDividerProps>} \cr}

\item{onRenderHeader}{\verb{IRenderFunction<IDetailsGroupDividerProps>} \cr}

\item{ariaLabelForToggleAllGroupsButton}{\code{string} \cr ariaLabel for expand/collapse group button}

\item{columnReorderProps}{\code{IColumnReorderHeaderProps} \cr Column reordering options}

\item{isAllCollapsed}{\code{boolean} \cr Whether or not all is collapsed}

\item{onColumnAutoResized}{\verb{(column: IColumn, columnIndex: number) => void} \cr Callback for when column is automatically resized}

\item{onColumnClick}{\verb{(ev: React.MouseEvent<HTMLElement>, column: IColumn) => void} \cr Callback for when the column is clicked}

\item{onColumnContextMenu}{\verb{(column: IColumn, ev: React.MouseEvent<HTMLElement>) => void} \cr Callback for when the column needs to show a context menu}

\item{onColumnIsSizingChanged}{\verb{(column: IColumn, isSizing: boolean) => void} \cr Callback for when column sizing has changed}

\item{onColumnResized}{\verb{(column: IColumn, newWidth: number, columnIndex: number) => void} \cr Callback for when column is resized}

\item{onRenderColumnHeaderTooltip}{\verb{IRenderFunction<IDetailsColumnRenderTooltipProps>} \cr Callback to render a tooltip for the column header}

\item{onToggleCollapseAll}{\verb{(isAllCollapsed: boolean) => void} \cr Callback for when collapse all is toggled}

\item{selectAllVisibility}{\code{SelectAllVisibility} \cr Select all button visibility}

\item{groupNestingDepth}{\code{number} \cr Nesting depth of a grouping}

\item{rowWidth}{\code{number} \cr Minimum width of the row.}

\item{ariaLabel}{\code{string} \cr Accessible label describing or summarizing the list.}

\item{ariaLabelForGrid}{\code{string} \cr Accessible label for the grid within the list.}

\item{ariaLabelForListHeader}{\code{string} \cr Accessible label for the list header.}

\item{ariaLabelForSelectAllCheckbox}{\code{string} \cr Accessible label for the select all checkbox.}

\item{ariaLabelForSelectionColumn}{\code{string} \cr Accessible label for the name of the selection column.}

\item{checkboxVisibility}{\code{CheckboxVisibility} \cr Controls the visibility of selection check box.}

\item{columnReorderOptions}{\code{IColumnReorderOptions} \cr Options for column reordering using drag and drop.}

\item{constrainMode}{\code{ConstrainMode} \cr Controls how the list contrains overflow.}

\item{disableSelectionZone}{\code{boolean} \cr Whether to disable the built-in SelectionZone, so the host component can provide its own.}

\item{enterModalSelectionOnTouch}{\code{boolean} \cr Whether the selection zone should enter modal state on touch.}

\item{getCellValueKey}{\verb{(item?: any, index?: number, column?: IColumn) => string} \cr If provided, will be the "default" item column cell value return. A column's \code{getValueKey} can override \code{getCellValueKey}.}

\item{getGroupHeight}{\code{IGroupedListProps['getGroupHeight']} \cr Callback to override default group height calculation used by list virtualization.}

\item{getKey}{\verb{(item: any, index?: number) => string} \cr Callback to get the item key, to be used in the selection and on render. Must be provided if sorting or filtering is enabled.}

\item{groupProps}{\code{IDetailsGroupRenderProps} \cr Override properties to render groups.}

\item{groups}{\code{IGroup[]} \cr Grouping instructions.}

\item{indentWidth}{\code{number} \cr Override for the indent width used for group nesting.}

\item{initialFocusedIndex}{\code{number} \cr Default index to set focus to once the items have rendered and the index exists.}

\item{isHeaderVisible}{\code{boolean} \cr Controls the visibility of the header.}

\item{isPlaceholderData}{\code{boolean} \cr Set this to true to indicate that the items being displayed are placeholder data.}

\item{items}{\code{any[]} \cr The items to render.}

\item{layoutMode}{\code{DetailsListLayoutMode} \cr Controls how the columns are adjusted.}

\item{listProps}{\code{IListProps} \cr Properties to pass through to the List components being rendered.}

\item{minimumPixelsForDrag}{\code{number} \cr The minimum mouse move distance to interpret the action as drag event.}

\item{onActiveItemChanged}{\verb{(item?: any, index?: number, ev?: React.FocusEvent<HTMLElement>) => void} \cr Callback for when an item in the list becomes active by clicking anywhere inside the row or navigating to it with the keyboard.}

\item{onColumnHeaderClick}{\verb{(ev?: React.MouseEvent<HTMLElement>, column?: IColumn) => void} \cr Callback for when the user clicks on the column header.}

\item{onColumnHeaderContextMenu}{\verb{(column?: IColumn, ev?: React.MouseEvent<HTMLElement>) => void} \cr Callback for when the user asks for a contextual menu (usually via right click) from a column header.}

\item{onColumnResize}{\verb{(column?: IColumn, newWidth?: number, columnIndex?: number) => void} \cr Callback fired on column resize}

\item{onDidUpdate}{\verb{(detailsList?: DetailsListBase) => void} \cr Callback for when the list has been updated. Useful for telemetry tracking externally.}

\item{onItemContextMenu}{\verb{(item?: any, index?: number, ev?: Event) => void | boolean} \cr Callback for when the context menu of an item has been accessed. If undefined or false is returned, \code{ev.preventDefault()} will be called.}

\item{onItemInvoked}{\verb{(item?: any, index?: number, ev?: Event) => void} \cr Callback for when a given row has been invoked (by pressing enter while it is selected.)}

\item{onRenderCheckbox}{\verb{IRenderFunction<IDetailsListCheckboxProps>} \cr If provided, can be used to render a custom checkbox.}

\item{onRenderDetailsFooter}{\verb{IRenderFunction<IDetailsFooterProps>} \cr An override to render the details footer.}

\item{onRenderDetailsHeader}{\verb{IRenderFunction<IDetailsHeaderProps>} \cr An override to render the details header.}

\item{onRenderItemColumn}{\verb{(item?: any, index?: number, column?: IColumn) => React.ReactNode} \cr If provided, will be the "default" item column renderer method. This affects cells within the rows, not the rows themselves. If a column definition provides its own \code{onRender} method, that will be used instead of this.}

\item{onRenderMissingItem}{\verb{(index?: number, rowProps?: IDetailsRowProps) => React.ReactNode} \cr Callback for what to render when the item is missing.}

\item{onRenderRow}{\verb{IRenderFunction<IDetailsRowProps>} \cr Callback to override the default row rendering.}

\item{onRowDidMount}{\verb{(item?: any, index?: number) => void} \cr Callback for when a given row has been mounted. Useful for identifying when a row has been rendered on the page.}

\item{onRowWillUnmount}{\verb{(item?: any, index?: number) => void} \cr Callback for when a given row has been unmounted. Useful for identifying when a row has been removed from the page.}

\item{onShouldVirtualize}{\verb{(props: IListProps) => boolean} \cr Callback to determine whether the list should be rendered in full, or virtualized.

Virtualization will add and remove pages of items as the user scrolls them into the visible range. This benefits larger list scenarios by reducing the DOM on the screen, but can negatively affect performance for smaller lists.

The default implementation will virtualize when this callback is not provided.}

\item{rowElementEventMap}{\verb{\{ eventName: string; callback: (context: IDragDropContext, event?: any) => void; \}[]} \cr Event names and corresponding callbacks that will be registered to rendered row elements.}

\item{selectionPreservedOnEmptyClick}{\code{boolean} \cr By default, selection is cleared when clicking on an empty (non-focusable) section of the screen. Setting this value to true overrides that behavior and maintains selection.}

\item{selectionZoneProps}{\code{ISelectionZoneProps} \cr Additional props to pass through to the SelectionZone created by default.}

\item{setKey}{\code{string} \cr A key that uniquely identifies the given items. If provided, the selection will be reset when the key changes.}

\item{shouldApplyApplicationRole}{\code{boolean} \cr Whether the role \code{application} should be applied to the list.}

\item{usePageCache}{\code{boolean} \cr Whether to enable render page caching. This is an experimental performance optimization that is off by default.}

\item{viewport}{\code{IViewport} \cr Viewport info, provided by the \code{withViewport} decorator.}

\item{cellsByColumn}{\verb{\{ [columnKey: string]: React.ReactNode; \}} \cr Optional pre-rendered content per column. Preferred over onRender or onRenderItemColumn if provided.}

\item{checkboxCellClassName}{\code{string} \cr Class name for the checkbox cell}

\item{checkButtonAriaLabel}{\code{string} \cr Check button's aria label}

\item{collapseAllVisibility}{\code{CollapseAllVisibility} \cr Collapse all visibility}

\item{componentRef}{\verb{IRefObject<IDetailsRow>} \cr Ref of the component}

\item{dragDropEvents}{\code{IDragDropEvents} \cr Handling drag and drop events}

\item{dragDropHelper}{\code{IDragDropHelper} \cr Helper for the drag and drop}

\item{eventsToRegister}{\verb{\{ eventName: string; callback: (item?: any, index?: number, event?: any) => void; \}[]} \cr A list of events to register}

\item{getRowAriaDescribedBy}{\verb{(item: any) => string} \cr Callback for getting the row aria-describedby}

\item{getRowAriaLabel}{\verb{(item: any) => string} \cr Callback for getting the row aria label}

\item{onDidMount}{\verb{(row?: DetailsRowBase) => void} \cr Callback for did mount for parent}

\item{onRenderCheck}{\verb{(props: IDetailsRowCheckProps) => JSX.Element} \cr Callback for rendering a checkbox}

\item{onWillUnmount}{\verb{(row?: DetailsRowBase) => void} \cr Callback for will mount for parent}

\item{rowFieldsAs}{\verb{React.ComponentType<IDetailsRowFieldsProps>} \cr DOM element into which to render row field}

\item{useReducedRowRenderer}{\code{boolean} \cr Rerender DetailsRow only when props changed. Might cause regression when depending on external updates.}

\item{anySelected}{\code{boolean} \cr Is any selected - also true for isSelectionModal}

\item{canSelect}{\code{boolean} \cr Can this checkbox be selectable}

\item{checkClassName}{\code{string} \cr The classname to be passed down to Check component}

\item{className}{\code{string} \cr Optional className to attach to the slider root element.}

\item{isHeader}{\code{boolean} \cr Is the check part of the header in a DetailsList}

\item{isVisible}{\code{boolean} \cr Whether or not this checkbox is visible}

\item{onRenderDetailsCheckbox}{\verb{IRenderFunction<IDetailsCheckboxProps>} \cr If provided, can be used to render a custom checkbox}

\item{selected}{\code{boolean} \cr Whether or not this check is selected}

\item{theme}{\code{ITheme} \cr Theme provided by High-Order Component.}

\item{useFastIcons}{\code{boolean} \cr Whether to use fast icon and check components. The icons can't be targeted by customization but are still customizable via class names.}

\item{cellStyleProps}{\code{ICellStyleProps} \cr Style properties to customize cell render output.}

\item{columnStartIndex}{\code{number} \cr Index to start for the column}

\item{compact}{\code{boolean} \cr whether to render as a compact field}

\item{enableUpdateAnimations}{\code{boolean} \cr}

\item{item}{\code{any} \cr Data source for this component}

\item{itemIndex}{\code{number} \cr The item index of the collection for the DetailsList}

\item{rowClassNames}{\verb{\{ [k in keyof Pick<IDetailsRowStyles, 'isMultiline' | 'isRowHeader' | 'cell' | 'cellAnimation' | 'cellPadded' | 'cellUnpadded' | 'fields'>]: string; \}} \cr Subset of classnames currently generated in DetailsRow that are used within DetailsRowFields.}

\item{columns}{\code{IColumn[]} \cr Column metadata}

\item{selection}{\code{ISelection} \cr Selection from utilities}

\item{selectionMode}{\code{SelectionMode} \cr Selection mode}

\item{ariaLabelForShimmer}{\code{string} \cr Aria label for shimmer. Set on grid while shimmer is enabled.}

\item{detailsListStyles}{\code{IDetailsListProps['styles']} \cr DetailsList styles to pass through.}

\item{enableShimmer}{\code{boolean} \cr Boolean flag to control when to render placeholders vs real items. It's up to the consumer app to know when fetching of the data is done to toggle this prop.}

\item{onRenderCustomPlaceholder}{\verb{(rowProps: IDetailsRowProps, index?: number, defaultRender?: (props: IDetailsRowProps) => React.ReactNode) => React.ReactNode} \cr Custom placeholder renderer to be used when in need to override the default placeholder of a DetailsRow. \code{rowProps} argument is passed to leverage the calculated column measurements done by DetailsList or you can use the optional arguments of item \code{index} and \code{defaultRender} to execute additional logic before rendering the default placeholder.}

\item{removeFadingOverlay}{\code{boolean} \cr Determines whether to remove a fading out to bottom overlay over the shimmering items used to further emphasize the unknown number of items that will be fetched.}

\item{shimmerLines}{\code{number} \cr Number of shimmer placeholder lines to render.}

\item{shimmerOverlayStyles}{\verb{IStyleFunctionOrObject<IShimmeredDetailsListStyleProps, IShimmeredDetailsListStyles>} \cr Custom styles to override the styles specific to the ShimmeredDetailsList root area.}

\item{styles}{\verb{IStyleFunctionOrObject<IShimmeredDetailsListStyleProps, IShimmeredDetailsListStyles>} \cr Custom styles to override the styles specific to the ShimmeredDetailsList root area.}

\item{skipViewportMeasures}{\code{boolean} \cr Whether or not to use ResizeObserver (if available) to detect and measure viewport on 'resize' events.

Falls back to window 'resize' event.}
}
\description{
A details list (\code{DetailsList}) is a robust way to display an information-rich collection of items, and allow people to sort, group, and filter the content. Use a details list when information density is critical.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/DetailsList}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\section{Best practices}{
\subsection{Layout}{
\itemize{
\item List items are composed of selection, icon, and name columns at minimum. You can include other columns, such as date modified, or any other metadata field associated with the collection.
\item Avoid using file type icon overlays to denote status of a file as it can make the entire icon unclear.
\item If there are multiple lines of text in a column, consider the variable row height variant.
\item Give columns ample default width to display information.
}
}

\subsection{Content}{
\itemize{
\item Use sentence-style capitalization for column headers—only capitalize the first word. For more info, see \href{https://docs.microsoft.com/style-guide/capitalization}{Capitalization} in the Microsoft Writing Style Guide.
}
}

\subsection{FAQ}{
\subsection{My scrollable content isn't updating on scroll. What should I do?}{

Add the \code{data-is-scrollable="true"} attribute to your scrollable element containing the DetailsList.

By default, the List used within DetailsList will use the \code{body} element as the scrollable element. If you contain the List within a scrollable \code{div} using \code{overflow: auto} or \code{scroll}, the List needs to listen for scroll events on that element instead. On initialization, the List will traverse up the DOM looking for the first element with the \code{data-is-scrollable} attribute to know which element to listen to for knowing when to re-evaulate the visible window.
}

\subsection{My List is not re-rendering when I mutate its items. What should I do?}{

To determine if the List within DetailsList should re-render its contents, the component performs a referential equality check within its \code{shouldComponentUpdate} method. This is done to minimize the performance overhead associated with re-rendering the virtualized List pages, as recommended by the \href{https://reactjs.org/docs/optimizing-performance.html#the-power-of-not-mutating-data}{React documentation}.

As a result of this implementation, the inner List will not determine it should re-render if the array values are mutated.
To avoid this problem, we recommend re-creating the items array backing the DetailsList by using a method such as \code{Array.prototype.concat} or ES6 spread syntax shown below:\if{html}{\out{<div class="tsx">}}\preformatted{public appendItems(): void \{
  const \{ items \} = this.state;

  this.setState(\{
    items: [...items, ...['Foo', 'Bar']]
  \})
\}

public render(): JSX.Element \{
  const \{ items \} = this.state;

  return <DetailsList items=\{items\} />;
\}
}\if{html}{\out{</div>}}

By re-creating the items array without mutating the values, the inner List will correctly determine its contents have changed and it should then re-render with the new values.
}

}
}

\examples{
library(shiny.fluent)

if (interactive()) {
  items <- list(
    list(key = "1", name = "Mark", surname = "Swanson"),
    list(key = "2", name = "Josh", surname = "Johnson")
  )

  columns <- list(
    list(key = "name", fieldName = "name", name = "Name"),
    list(key = "surname", fieldName = "surname", name = "Surname")
  )

  shinyApp(
    ui = withReact(
      DetailsList(items = items, columns = columns)
    ),
    server = function(input, output) {
    }
  )
}
}
