% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{CompactPeoplePicker}
\alias{CompactPeoplePicker}
\alias{NormalPeoplePicker}
\alias{PeoplePicker}
\title{PeoplePicker}
\usage{
CompactPeoplePicker(...)

NormalPeoplePicker(...)
}
\arguments{
\item{ValidationState}{\code{ValidationState} \cr}

\item{className}{\code{string} \cr Additional CSS class(es) to apply to the PeoplePickerItem root element.}

\item{theme}{\code{ITheme} \cr Theme provided by High-Order Component.}

\item{compact}{\code{boolean} \cr Flag that controls whether each suggested PeoplePicker item (Persona) is rendered with or without secondary text for compact look.}

\item{personaProps}{\code{IPersonaProps} \cr Persona props for each suggested for picking PeoplePicker item.}

\item{styles}{\verb{IStyleFunctionOrObject<IPeoplePickerItemSuggestionStyleProps, IPeoplePickerItemSuggestionStyles>} \cr Call to provide customized styling that will layer on top of the variant rules.}

\item{suggestionsProps}{\code{IBasePickerSuggestionsProps} \cr General common props for all PeoplePicker items suggestions.}
}
\description{
The people picker (\code{PeoplePicker}) is used to select one or more entities, such as people or groups, from a list. It makes composing an email to someone, or adding them to a group, easy if you don’t know their full name or email address.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/PeoplePicker}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\section{Best practices}{
\subsection{Layout}{
\itemize{
\item Use the people picker to add someone to the To line of an email, or to add someone to a list.
\item Use the \verb{MemberList PeoplePicker} to display selections below the input field.
}
}
}

