% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R,
%   R/inputs.R
\name{Rating}
\alias{Rating}
\alias{Rating.shinyInput}
\alias{updateRating.shinyInput}
\title{Rating}
\usage{
Rating(...)

Rating.shinyInput(inputId, ..., value = defaultValue)

updateRating.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{allowZeroStars}{\code{boolean} \cr Allow the rating value to be set to 0 instead of a minimum of 1.}

\item{ariaLabelFormat}{\code{string} \cr Optional label format for a rating star that will be read by screen readers. Can be used like "\{0\} of \{1\} stars selected", where \{0\} will be substituted by the current rating and \{1\} will be substituted by the max rating.}

\item{ariaLabelId}{\code{string} \cr Deprecated: Optional id of label describing this instance of Rating.}

\item{componentRef}{\verb{IRefObject<IRating>} \cr Optional callback to access the IRating interface. Use this instead of ref for accessing the public methods and properties of the component.}

\item{getAriaLabel}{\verb{(rating: number, max: number) => string} \cr}

\item{icon}{\code{string} \cr Custom icon}

\item{max}{\code{number} \cr Maximum rating, defaults to 5, has to be \>= min}

\item{min}{\code{number} \cr Minimum rating, defaults to 1, has to be \>= 0}

\item{onChange}{\verb{(event: React.FocusEvent<HTMLElement>, rating?: number) => void} \cr Callback issued when the rating changes.}

\item{onChanged}{\verb{(rating: number) => void} \cr}

\item{rating}{\code{number} \cr Selected rating, has to be an integer between min and max}

\item{readOnly}{\code{boolean} \cr Optional flag to mark rating control as readOnly}

\item{size}{\code{RatingSize} \cr Size of rating, defaults to small}

\item{styles}{\verb{IStyleFunctionOrObject<IRatingStyleProps, IRatingStyles>} \cr Call to provide customized styling that will layer on top of the variant rules.}

\item{theme}{\code{ITheme} \cr Theme (provided through customization.)}

\item{unselectedIcon}{\code{string} \cr Custom icon for unselected rating elements.}
}
\description{
Ratings show people’s opinions of a product, helping others make more informed purchasing decisions. People can also rate products they’ve purchased.

For more details and examples visit the official \href{https://developer.microsoft.com/en-us/fluentui#/controls/web/Rating}{docs}.
The R package can not handle each and every case, so for advanced use cases
you need to work using the original docs to achieve the desired result.
}
\section{Best practices}{
\subsection{Layout}{
\itemize{
\item Make it clear which item the rating pertains to by making sure the layout and grouping are clear when several items are on the page.
\item Don't use the rating component for data that has a continuous range, such as the brightness of a photo. Instead, use a slider.
}
}

\subsection{Content}{
\itemize{
\item Use a five-star rating system.
\item Use sentence-style capitalization—only capitalize the first word. For more info, see \href{https://docs.microsoft.com/style-guide/capitalization}{Capitalization} in the Microsoft Writing Style Guide.
}
}
}

\examples{
library(shiny.fluent)

if (interactive()) {
  shinyApp(
    ui = div(
      Rating.shinyInput("rating", value = 2),
      textOutput("ratingValue")
    ),
    server = function(input, output) {
      output$ratingValue <- renderText({
        sprintf("Value: \%s", input$rating)
      })
    }
  )
}
}
